requirements.txt
pydantic>=2.7,<3
python-dotenv>=1.0,<2
faster-whisper>=1.0,<2
requests>=2.31,<3
playwright>=1.44,<2
python-json-logger>=2.0,<3
sounddevice>=0.4.7,<1
numpy>=1.26,<3
scipy>=1.11,<2
pytest>=8.0,<9
